requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.6.0
python-dotenv==0.15.0
toml==0.10.2
Werkzeug==1.0.1
